Editing part D

+++ src/main/java/D/Service/EmployeeService.java
@@ -2,5 +2,5 @@ package D.Service;
 
 import D.Entities.Employee;
 
-public interface EmployeeService extends BaseService<Employee,Integer> {
+public interface EmployeeService extends BaseService<Employee,Long> {
 }
